1_ChatbotUsingLLM/requirements.txt
python-dotenv
aiogram==3.*
openai